+++ MonopolyProject/Models/Space.cs
@@ -4,12 +4,12 @@ namespace MonopolyProject.Models
         {
             public string Name { get; set; }
             public SpaceType Type { get; set; }
-            public int Price { get; set; }
+            public double Price { get; set; }
             public ColorType ColorGroup { get; set; } 
             public Player Owner { get; set; }
             public int HouseCount { get; set; }
 
-            public Space(string name, SpaceType type, int price, ColorType colorGroup)
+            public Space(string name, SpaceType type, double price, ColorType colorGroup)
             {
                 Name = name;
                 Type = type;
@@ -20,17 +20,17 @@ namespace MonopolyProject.Models
             }
 
             // Method to calculate rent based on house count
-            public int CalculateRent()
+            public double CalculateRent()
             {
                 double baseRent = Price * 0.25;
                 double houseRent = Price * 0.75 * HouseCount;
-                return (int)(baseRent + houseRent);
+                return baseRent + houseRent;
             }
 
             // Method to get the cost of building a house
-            public int HouseCost()
+            public double HouseCost()
             {
-                return (int)(Price * 0.6);
+                return Price * 0.6;
             }
         }
 }
